Update Terraform repl package to bring in better formatting
The repl package in Terraform core was updated to include heredoc formatting
for multiline strings. The interface for this package changed to use cty.Value
rather than map[string]interface{} so I had to bring in cty as a dependency
and update the code that munges the manifest after it has been decoded to use it.